module mailattach moved to src (from extensions) + lic info added
